rename type_prefix => register_type_prefix
