+++ test/registertypeprefix/Makefile
@@ -1,0 +1,2 @@
+regenerate:
+	(protoc-min-version --version="3.0.0" --proto_path=../../../../../:../../protobuf/:. --gogofast_out=. registertypeprefix.proto)
