+++ test/registertypeprefix/registertypeprefix.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 
-option (gogoproto.type_prefix) = "FooBar";
+option (gogoproto.register_type_prefix) = "FooBar";
 
 message Foo {
     string bar = 1;
