+++ test/registertypeprefix/registertypeprefix_test.go
@@ -1,4 +1,4 @@
-package typeprefix
+package registertypeprefix
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 )
 
-func TestTypePrefix(t *testing.T) {
+func TestRegisterTypePrefix(t *testing.T) {
 	wantedName := "FooBar.Foo"
 	if proto.MessageName(&Foo{}) != wantedName {
 		t.Fatalf("proto.MessageName(&Foo{}) != %s", wantedName)
