--- test/typeprefix/Makefile
@@ -1,2 +1,0 @@
-regenerate:
-	(protoc-min-version --version="3.0.0" --proto_path=../../../../../:../../protobuf/:. --gogofast_out=. typeprefix.proto)
